poll_oneoff: add wasi end-to-end tests
