+++ imports/wasi_snapshot_preview1/wasi_test.go
@@ -4,7 +4,9 @@ import (
 	"bytes"
 	"context"
 	_ "embed"
+	"io/fs"
 	"testing"
+	"time"
 
 	"github.com/tetratelabs/wazero"
 	"github.com/tetratelabs/wazero/api"
@@ -147,3 +149,30 @@ func requireErrnoResult(t *testing.T, expectedErrno wasip1.Errno, mod api.Closer
 	errno := wasip1.Errno(results[0])
 	require.Equal(t, expectedErrno, errno, "want %s but have %s", wasip1.ErrnoName(expectedErrno), wasip1.ErrnoName(errno))
 }
+
+func newBlockingReader(t *testing.T) blockingReader {
+	timeout, cancelFunc := context.WithTimeout(testCtx, 5*time.Second)
+	t.Cleanup(cancelFunc)
+	return blockingReader{ctx: timeout}
+}
+
+// blockingReader is a reader that never terminates its read
+// unless the embedded context is Done()
+type blockingReader struct {
+	ctx context.Context
+}
+
+func (b blockingReader) Read(p []byte) (n int, err error) {
+	<-b.ctx.Done()
+	return 0, nil
+}
+
+// stdinFileInfo implements fs.FileInfo: it is only representing the mode because it is always stdin
+type stdinFileInfo uint32
+
+func (stdinFileInfo) Name() string        { return "stdin" }
+func (stdinFileInfo) Size() int64         { return 0 }
+func (s stdinFileInfo) Mode() fs.FileMode { return fs.FileMode(s) }
+func (stdinFileInfo) ModTime() time.Time  { return time.Unix(0, 0) }
+func (stdinFileInfo) IsDir() bool         { return false }
+func (stdinFileInfo) Sys() interface{}    { return nil }
